refactor: separate generic window files
